# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Classroom Supplies Usage & Replenishment Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}/reset-password:
    post:
      operationId: resetPassword
      tags: [users]
      summary: Reset a teacher's password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Success
  /users/{id}/status:
    post:
      operationId: setUserStatus
      tags: [users]
      summary: Enable or disable a teacher's account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserStatusInput"
      responses:
        "200":
          description: Success
components:
  schemas:
...
    ResetPasswordInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 6
    SetUserStatusInput:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
...